Restrukturering etter diskusjon
