+++ src/main/kotlin/no/nav/familie/ks/sak/api/ekstern/EksternKlageController.kt
@@ -1,10 +1,14 @@
-package no.nav.familie.ks.sak.ekstern
+package no.nav.familie.ks.sak.api.ekstern
 
 import no.nav.familie.kontrakter.felles.Ressurs
+import no.nav.familie.kontrakter.felles.klage.FagsystemVedtak
 import no.nav.familie.kontrakter.felles.klage.KanOppretteRevurderingResponse
 import no.nav.familie.kontrakter.felles.klage.OpprettRevurderingResponse
 import no.nav.familie.ks.sak.common.exception.Feil
 import no.nav.familie.ks.sak.config.BehandlerRolle
+import no.nav.familie.ks.sak.kjerne.behandling.OpprettBehandlingService
+import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingÅrsak
+import no.nav.familie.ks.sak.kjerne.klage.KlageService
 import no.nav.familie.ks.sak.sikkerhet.AuditLoggerEvent
 import no.nav.familie.ks.sak.sikkerhet.SikkerhetContext
 import no.nav.familie.ks.sak.sikkerhet.TilgangService
@@ -16,22 +20,24 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
+// Kalles av familie-klage
 @RestController
 @RequestMapping(
-    path = ["/api/ekstern/behandling"],
+    path = ["/api/ekstern"],
     produces = [MediaType.APPLICATION_JSON_VALUE]
 )
 @ProtectedWithClaims(issuer = "azuread")
-class EksternBehandlingController(
+class EksternKlageController(
     private val tilgangService: TilgangService,
-    private val eksternBehandlingService: EksternBehandlingService
+    private val opprettBehandlingService: OpprettBehandlingService,
+    private val klageService: KlageService
 ) {
 
-    @GetMapping("kan-opprette-revurdering/{fagsakId}")
-    fun kanOppretteRevurdering(@PathVariable fagsakId: Long): Ressurs<KanOppretteRevurderingResponse> {
+    @GetMapping("fagsaker/{fagsakId}/kan-opprette-revurdering")
+    fun kanOppretteKlageRevurdering(@PathVariable fagsakId: Long): Ressurs<KanOppretteRevurderingResponse> {
         tilgangService.validerTilgangTilHandlingOgFagsak(
             fagsakId = fagsakId,
-            handling = "Kan opprette revurdering fra klage på fagsak=$fagsakId",
+            handling = "Kan opprette revurdering på fagsak=$fagsakId",
             event = AuditLoggerEvent.CREATE,
             minimumBehandlerRolle = BehandlerRolle.SAKSBEHANDLER
         )
@@ -40,10 +46,10 @@ class EksternBehandlingController(
             throw Feil("Kallet utføres ikke av en autorisert klient")
         }
 
-        return Ressurs.success(eksternBehandlingService.kanOppretteRevurdering(fagsakId))
+        return Ressurs.success(opprettBehandlingService.kanOppretteRevurdering(fagsakId))
     }
 
-    @PostMapping("opprett-revurdering-klage/{fagsakId}")
+    @PostMapping("fagsaker/{fagsakId}/opprett-revurdering-klage/")
     fun opprettRevurderingKlage(@PathVariable fagsakId: Long): Ressurs<OpprettRevurderingResponse> {
         tilgangService.validerTilgangTilHandlingOgFagsak(
             fagsakId = fagsakId,
@@ -55,6 +61,21 @@ class EksternBehandlingController(
         if (!SikkerhetContext.kallKommerFraKlage()) {
             throw Feil("Kallet utføres ikke av en autorisert klient")
         }
-        return Ressurs.success(eksternBehandlingService.opprettRevurderingKlage(fagsakId))
+        return Ressurs.success(opprettBehandlingService.prøvÅOppretteRevurdering(fagsakId, BehandlingÅrsak.KLAGE))
+    }
+
+    @GetMapping("fagsaker/{fagsakId}/vedtak")
+    @ProtectedWithClaims(issuer = "azuread")
+    fun hentVedtak(@PathVariable fagsakId: Long): Ressurs<List<FagsystemVedtak>> {
+        if (!SikkerhetContext.erMaskinTilMaskinToken()) {
+            tilgangService.validerTilgangTilHandlingOgFagsak(
+                fagsakId = fagsakId,
+                handling = "Kan hente vedtak på fagsak=$fagsakId",
+                event = AuditLoggerEvent.ACCESS,
+                minimumBehandlerRolle = BehandlerRolle.SAKSBEHANDLER
+            )
+        }
+
+        return Ressurs.success(klageService.hentFagsystemVedtak(fagsakId))
     }
 }
